move Cli to separate file
Just a straight move, no logic change at all.
